+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayViewModel.kt
@@ -12,7 +12,6 @@ import com.rakuten.tech.mobile.miniapp.MiniAppDisplay
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
 import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -65,15 +64,4 @@ class MiniAppDisplayViewModel constructor(
             miniAppDisplay.navigateBackward()
         else
             false
-
-    private fun displayPermissions(
-        miniAppId: String,
-        permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>
-    ) {
-        if (::miniAppDisplay.isInitialized)
-            miniAppDisplay.showCustomPermissionWindow(
-                miniAppId,
-                permissionsWithDescription
-            )
-    }
 }